requirements.txt
pip
Flask
cassandra-driver
request
requests
requests_cache
